+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/TruffleFile.java
@@ -98,8 +98,10 @@ import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.impl.TruffleLocator;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.util.List;
 import java.util.ServiceLoader;
 import java.util.function.Supplier;
+import org.graalvm.polyglot.TypeLiteral;
 
 /**
  * An abstract representation of a file used by Truffle languages.
@@ -107,6 +109,143 @@ import java.util.function.Supplier;
  * @since 1.0
  */
 public final class TruffleFile {
+
+    /**
+     * The file's last modified time. Supported by all filesystems.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<FileTime> LAST_MODIFIED_TIME = new AttributeDescriptor<>(AttributeGroup.BASIC, "lastModifiedTime", FileTime.class);
+
+    /**
+     * The file's last access time. Supported by all filesystems.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<FileTime> LAST_ACCESS_TIME = new AttributeDescriptor<>(AttributeGroup.BASIC, "lastAccessTime", FileTime.class);
+
+    /**
+     * The file's creation time. Supported by all filesystems.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<FileTime> CREATION_TIME = new AttributeDescriptor<>(AttributeGroup.BASIC, "creationTime", FileTime.class);
+
+    /**
+     * Represents the file a regular file. Supported by all filesystems.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Boolean> IS_REGULAR_FILE = new AttributeDescriptor<>(AttributeGroup.BASIC, "isRegularFile", Boolean.class);
+
+    /**
+     * Represents the file a directory. Supported by all filesystems.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Boolean> IS_DIRECTORY = new AttributeDescriptor<>(AttributeGroup.BASIC, "isDirectory", Boolean.class);
+
+    /**
+     * Represents the file a symbolic link. Supported by all filesystems.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Boolean> IS_SYMBOLIC_LINK = new AttributeDescriptor<>(AttributeGroup.BASIC, "isSymbolicLink", Boolean.class);
+
+    /**
+     * Represents the file a special file (device, named pipe). Supported by all filesystems.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Boolean> IS_OTHER = new AttributeDescriptor<>(AttributeGroup.BASIC, "isOther", Boolean.class);
+
+    /**
+     * The file's size in bytes. Supported by all filesystems.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Long> SIZE = new AttributeDescriptor<>(AttributeGroup.BASIC, "size", Long.class);
+
+    /**
+     * The owner of the file. Supported only by UNIX native filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<UserPrincipal> UNIX_OWNER = new AttributeDescriptor<>(AttributeGroup.POSIX, "owner", UserPrincipal.class);
+
+    /**
+     * The group owner of the file. Supported only by UNIX native filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<GroupPrincipal> UNIX_GROUP = new AttributeDescriptor<>(AttributeGroup.POSIX, "group", GroupPrincipal.class);
+
+    /**
+     * The file's Posix permissions. Supported only by UNIX native filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Set<PosixFilePermission>> UNIX_PERMISSIONS = new AttributeDescriptor<>(AttributeGroup.POSIX, "permissions",
+                    new TypeLiteral<Set<PosixFilePermission>>() {
+                    });
+
+    /**
+     * The file's mode containing the protection and file type bits. Supported only by UNIX native
+     * filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Integer> UNIX_MODE = new AttributeDescriptor<>(AttributeGroup.UNIX, "mode", Integer.class);
+
+    /**
+     * The file's inode number. Supported only by UNIX native filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Long> UNIX_INODE = new AttributeDescriptor<>(AttributeGroup.UNIX, "ino", Long.class);
+
+    /**
+     * The id of a device containing the file. Supported only by UNIX native filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Long> UNIX_DEV = new AttributeDescriptor<>(AttributeGroup.UNIX, "dev", Long.class);
+
+    /**
+     * The id of a device represented by the file. Supported only by UNIX native filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Long> UNIX_RDEV = new AttributeDescriptor<>(AttributeGroup.UNIX, "rdev", Long.class);
+
+    /**
+     * The number of hard links. Supported only by UNIX native filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Integer> UNIX_NLINK = new AttributeDescriptor<>(AttributeGroup.UNIX, "nlink", Integer.class);
+
+    /**
+     * The user id of file owner. Supported only by UNIX native filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Integer> UNIX_UID = new AttributeDescriptor<>(AttributeGroup.UNIX, "uid", Integer.class);
+
+    /**
+     * The group id of file owner. Supported only by UNIX native filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<Integer> UNIX_GID = new AttributeDescriptor<>(AttributeGroup.UNIX, "gid", Integer.class);
+
+    /**
+     * The file's last status change time. Supported only by UNIX native filesystem.
+     *
+     * @since 1.0
+     */
+    public static final AttributeDescriptor<FileTime> UNIX_CTIME = new AttributeDescriptor<>(AttributeGroup.UNIX, "ctime", FileTime.class);
+
     private static final int MAX_BUFFER_SIZE = Integer.MAX_VALUE - 8;
     private static final int BUFFER_SIZE = 8192;
     private static final boolean LEGACY_FILETYPEDETECTORS_DISABLED = Boolean.getBoolean("graalvm.legacy.filetypedetectors.disabled");
@@ -1488,6 +1627,173 @@ public final class TruffleFile {
         }
     }
 
+    private static final class AttributeGroup {
+
+        static final AttributeGroup BASIC = new AttributeGroup("basic", null);
+        static final AttributeGroup POSIX = new AttributeGroup("posix", BASIC);
+        static final AttributeGroup UNIX = new AttributeGroup("unix", POSIX);
+
+        final String name;
+        private final AttributeGroup parent;
+
+        AttributeGroup(String name, AttributeGroup parent) {
+            this.name = name;
+            this.parent = parent;
+        }
+
+        boolean contains(AttributeGroup other) {
+            if (name.equals(other.name)) {
+                return true;
+            }
+            if (parent != null) {
+                return parent.contains(other);
+            }
+            return false;
+        }
+
+        @Override
+        public String toString() {
+            return name;
+        }
+    }
+
+    /**
+     * Represents a file's attribute.
+     * <p>
+     * Not all attributes are supported by all filesystems. When the filesystem does not support the
+     * attribute the attribute read fails with {@link UnsupportedOperationException}.
+     * <p>
+     * Attributes can be read one by one by the
+     * {@link TruffleFile#getAttribute(com.oracle.truffle.api.TruffleFile.AttributeDescriptor, java.nio.file.LinkOption...)
+     * getAttribute} method or as a bulk read operation by
+     * {@link TruffleFile#getAttributes(java.util.Collection, java.nio.file.LinkOption...)
+     * getAttributes} method. When more attributes are needed the bulk read provides better
+     * performance.
+     *
+     * @since 1.0
+     */
+    public static final class AttributeDescriptor<T> {
+
+        final AttributeGroup group;
+        final String name;
+        final Class<T> clazz;
+
+        AttributeDescriptor(AttributeGroup group, String name, Class<T> clazz) {
+            this.group = group;
+            this.name = name;
+            this.clazz = clazz;
+        }
+
+        AttributeDescriptor(AttributeGroup group, String name, TypeLiteral<T> typeLiteral) {
+            this.group = group;
+            this.name = name;
+            this.clazz = typeLiteral.getRawType();
+        }
+
+        /**
+         *
+         * {@inheritDoc}
+         *
+         * @since 1.0
+         */
+        @Override
+        public String toString() {
+            return group + ":" + name;
+        }
+    }
+
+    /**
+     * A view over file's attributes values obtained by
+     * {@link TruffleFile#getAttributes(java.util.Collection, java.nio.file.LinkOption...)
+     * getAttributes}.
+     *
+     * @since 1.0
+     */
+    public static final class Attributes {
+
+        private final Map<String, Object> delegate;
+
+        Attributes(Map<String, Object> delegate) {
+            this.delegate = delegate;
+        }
+
+        /**
+         * Returns the attribute value.
+         *
+         * @param descriptor the attribute to obtain value for
+         * @return the attribute value
+         * @throws IllegalArgumentException if the view does not contain a value for given attribute
+         * @since 1.0
+         */
+        public <T> T get(AttributeDescriptor<T> descriptor) {
+            Object value = delegate.get(descriptor.name);
+            if (value != null) {
+                return descriptor.clazz.cast(value);
+            } else if (delegate.containsKey(descriptor.name)) {
+                return null;
+            } else {
+                throw new IllegalArgumentException("Attributes don't contain " + descriptor.toString());
+            }
+        }
+    }
+
+    /**
+     * Reads a single file's attribute.
+     *
+     * @param attribute the attribute to read
+     * @param linkOptions the options determining how the symbolic links should be handled
+     * @return the attribute value
+     * @throws IOException in case of IO error
+     * @throws UnsupportedOperationException when the filesystem does not support required
+     *             attribute.
+     * @throws SecurityException if the {@link FileSystem} denied the operation
+     * @since 1.0
+     */
+    public <T> T getAttribute(AttributeDescriptor<T> attribute, LinkOption... linkOptions) throws IOException {
+        try {
+            return getAttributeImpl(createAttributeString(attribute.group, Collections.singleton(attribute.name)), attribute.clazz, linkOptions);
+        } catch (IOException | UnsupportedOperationException | SecurityException e) {
+            throw e;
+        } catch (Throwable t) {
+            throw wrapHostException(t);
+        }
+    }
+
+    /**
+     * Reads file's attributes as a bulk operation.
+     *
+     * @param attributes the attributes to read
+     * @param linkOptions the options determining how the symbolic links should be handled
+     * @return the {@link Attributes attributes view}
+     * @throws IllegalArgumentException when no attributes are given
+     * @throws IOException in case of IO error
+     * @throws UnsupportedOperationException when the filesystem does not support some of the
+     *             required attributes.
+     * @throws SecurityException if the {@link FileSystem} denied the operation
+     * @since 1.0
+     */
+    public Attributes getAttributes(Collection<? extends AttributeDescriptor<?>> attributes, LinkOption... linkOptions) throws IOException {
+        if (attributes.isEmpty()) {
+            throw new IllegalArgumentException("No descriptors given.");
+        }
+        try {
+            AttributeGroup group = null;
+            List<String> attributeNames = new ArrayList<>();
+            for (AttributeDescriptor<?> descriptor : attributes) {
+                if (group == null || !group.contains(descriptor.group)) {
+                    group = descriptor.group;
+                }
+                attributeNames.add(descriptor.name);
+            }
+            Map<String, Object> map = fileSystemContext.fileSystem.readAttributes(normalizedPath, createAttributeString(group, attributeNames), linkOptions);
+            return new Attributes(map);
+        } catch (IOException | UnsupportedOperationException | SecurityException e) {
+            throw e;
+        } catch (Throwable t) {
+            throw wrapHostException(t);
+        }
+    }
+
     /**
      * A detector for finding {@link TruffleFile file}'s MIME type and encoding.
      *
@@ -1558,6 +1864,11 @@ public final class TruffleFile {
             }
             return filtered;
         }
+    }
+
+    private String createAttributeString(AttributeGroup group, Iterable<String> attributeNames) {
+        String joinedNames = String.join(",", attributeNames);
+        return group == AttributeGroup.BASIC ? joinedNames : group.name + ':' + joinedNames;
     }
 
     private boolean isNormalized() {